requirements.txt
boto3==1.22.4
botocore==1.25.4
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-auth==2.6.6
google-auth-oauthlib==0.5.1
gspread==5.3.2
idna==3.3
jmespath==1.0.0
more-itertools==8.12.0
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
